move some files around
