+++ src/components/modal/IO/exportButtons.tsx
@@ -1,8 +1,8 @@
 "use client"
 import Button from "@/components/toolBar/button";
 import { downloadTextAsFile } from "@/lib/missionIO/common";
+import { waypointTo_waypoints_file } from "@/lib/missionIO/waypointToFile";
 import { useWaypoints } from "@/util/context/WaypointContext";
-import { waypointTo_waypoints_file } from "@/util/waypointToFile";
 import { FaFileDownload } from "react-icons/fa";
 
 export default function DownloadButtons() {
