+++ src/lib/missionIO/waypointToFile.ts
